frontend/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.11.7
python-jose==3.3.0
python-multipart==0.0.6
sqlalchemy==2.0.25
numpy==2.2.6
opencv-python==4.12.0.88
pyserial==3.5
python-dotenv==1.0.0
confluent-kafka==2.3.0
aiosqlite==0.19.0
aiofiles==23.2.1
